Added Encoding and Decoding functions.

+++ src/main/java/compressor/HuffmanBaseNode.java
@@ -3,7 +3,7 @@ package src.main.java.compressor;
 /** Base node for Huffman tree */
 public interface HuffmanBaseNode {
     boolean isLeaf();
-    long getWeight();
+    int getWeight();
 
     HuffmanBaseNode getLeft();
     HuffmanBaseNode getRight();
